Modified the build file to access a local ftp server (IIS FTP Server running on localhost w/ anonymous authentication).
Modified the ftptask.cs file DIR routine to list the local directory instead of an invalid one.  

Using DirDetails on an invalid directory does not throw an exception but rather returns an empty listing.  Unfortunately, listing on an empty directory also returns an empty listing, so perhaps it is the chdir command after all to test for dir existance?


git-svn-id: svn://localhost:3690/nantftptask/trunk@28 d147aa11-f70d-cc43-991b-859307a355b7