Progress on the getfileset scanning routines.
RemotePath stil needs some work interpreting how it is used when it's FTPFile is null.

Currently, this set of files is all partially through the transition to enable remote scanning for the <get> include elements.

This revision is untested after modifications to prevent a nullreference exception when accessing RemoteFile.FullPath inside Clean(dir, dir)


git-svn-id: svn://localhost:3690/nantftptask/trunk@29 d147aa11-f70d-cc43-991b-859307a355b7