Renamed the task and type attributes to more consistant names (path-->dir).
Added support for a .ftp_password xml file with one root node <ftp password="pass" /> that gets tried for authentication before the console input feature is invoked.  This is provided as a way to save passwords outside of the build file but avoid re-entering them every time.

Updated the build file to correct the clean targets and a problem if the version attempts to read version info from a nonexistent assembly.  (this neads another rewrite, ideally, to handle more gracefully by building the assembly or running subwcrev or something)

Updated the prjx file to only import local copies of the lib assemblies and set localcopy to false for all other dirs.

Nant cannot be used to install a nant task while it's running (it's using the assembly in question!) so i've added an install and remove cmd that do brute-strength copy and rm.


git-svn-id: svn://localhost:3690/nantftptask/trunk@40 d147aa11-f70d-cc43-991b-859307a355b7