Whew!
Added msdn documentation and a build target 'msdndoc' to update it.

Added a 'doc' target that builds 'userdoc' and 'msdndoc'.

Added a NAnt.xsd schema generation task to the 'userdoc' target.

Enabled the <ascii-up>, etc and <script> commands, and updated them to use the edtFTPnet libraries.  Logging output is minimal at the info level, slightly better at the verbose level.

Added a NamespaceSummaries.xml file for use when generating 'msdndoc'.

Removed AssemblyInfo.cs, changed AssemblyInfo.subwcrev to AssemblyInfo.src, and changed the build targets to depend on AssemblyInfo.cs's existance as a flag.  If that file does not exist, it will be created from AssemblyInfo.src by a 'subwcrev' target.  


git-svn-id: svn://localhost:3690/nantftptask/trunk@41 d147aa11-f70d-cc43-991b-859307a355b7